outsourced GeometricUtil
