+++ app/lib/objectivemodeler/features/modeling/behavior/util/LabelLayoutUtil.js
@@ -1,4 +1,4 @@
-import {getAngle, getDistancePointPoint, rotateVector} from './GeometricUtil';
+import {getAngle, getDistancePointPoint, rotateVector} from '../../../../../frequentlyUsedFiles/features/modeling/behavior/util/GeometricUtil.js';
 
 import {getAttachment} from './LineAttachmentUtil';
